+++ CHANGELOG.md
@@ -47,6 +47,7 @@
   - `Uniq()` now supports tracing
   - `XargsBasename()` now supports tracing
   - `XargsCat()` now supports tracing
+  - `XargsDirname()` now supports tracing
 * Pipelines now set a context flag in their Pipe
   - this is used to tell sinks where to read from!